Finished Food Add and Chat UI

+++ app/src/main/java/com/unmsm/nutrihealth/data/model/Food.kt
@@ -3,7 +3,7 @@ package com.unmsm.nutrihealth.data.model
 data class Food(
     val name: String,
     val calories: Int,
-    val protein: Int,
-    val carbs: Int,
-    val fat: Int,
+    val protein: Float,
+    val carbs: Float,
+    val fat: Float,
 )
